requirements.txt
fastapi==0.116.1
langchain==0.3.26
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-mongodb==0.6.2
numpy==2.3.1
pydantic==2.11.7
pymongo==4.13.2
scikit-learn==1.7.0
uvicorn==0.35.0
python-dotenv==1.1.1
google-auth==2.35.0
